ScoreVersionDetailContent を分離
